Correct calculation of handle position
The previous algorithm wasn't properly placing the handle when the
scroll value approached 1.  In particular, it leaned heavily on CLAMP to
stop the handle from going offscreen.  We needed to rescale some
values to map scrollValue = 1 to the bottom of the handle,
scrollValue = 0 to the top of the handle, and scrollValue = 0.5 to the
middle.