Initial work to show an accessory view on scroll
This view can be used to show the current location in the window (e.g.:
to show the date of photos in a chronological list).  It is
automatically cenetered to the left of the handle (left by the hit
width, so that the view is placed to the left of the user's finger), and
hidden and shown when a user drags the handle.